Process: Make xdgTestHelper into a self-contained file
- Build as part of DarwinCompatibilityTests.

- Add TestBundle.xdgTestHelperURL() to return the URL of the
  executable.
